Fix alt text persistence
